apps/ai-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.2
python-multipart==0.0.20
pillow==11.1.0
numpy==2.2.2
opencv-python==4.10.0.84
torch==2.5.1
torchvision==0.20.1
transformers==4.48.2
ultralytics==8.3.62
requests==2.32.3
pydantic==2.10.6
scikit-image==0.24.0
